Added: fire module with function GetFire
